Added server-side authentication and a demo of it.
To do this, I added a Services folder inside of the management project,
I created a wrapper service for the HttpClient that adds the auth
header, I added an auth middleware that can validate the JWTs, and I
added a controller to the API that returns the authenticated user's
username if they authenticated correctly. This was tested in the
location/station page on the frontend.